move currencies and intervals to symbols
