+++ src/crypto_candlesticks/bitfinex_connector/__init__.py
@@ -1,15 +1,3 @@
 # -*- coding: utf-8 -*-
 """Crypto candlesticks api connector."""
-try:
-    from importlib.metadata import PackageNotFoundError, version
-except ImportError:  # pragma: no cover
-    from importlib_metadata import (  # type: ignore
-        PackageNotFoundError,
-        version,
-    )
-
-
-try:
-    __version__ = version(__name__)
-except PackageNotFoundError:  # pragma: no cover
-    __version__ = "unknown"
+from crypto_candlesticks.bitfinex_connector.connector import Connector
