+++ src/crypto_candlesticks/symbols/__init__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+"""Valid intervals and base currencies in Bitfinex."""
+from crypto_candlesticks.symbols.intervals import intervals
+from crypto_candlesticks.symbols.quote_currency import quote_currency
