+++ src/crypto_candlesticks/symbols/intervals.py
